requirements.txt
Pillow>=8.4.0
python-telegram-bot>=13.7